+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Send logout notifications to the Custom Metadata Handler in advanced settings. (#492)
 - Bugfix: Fire death notifications for Doom modifier in Fortis Colosseum. (#474)
 - Dev: Include player region and world to notification metadata for all notifiers. (#490)
 - Dev: Allow custom webhook handlers to use HTTP status code 307 and 308 to redirect requests. (#484)